fix bug with newly introduced non-null api res
Exercise data in Records is denormalized. On GET calls, mongo does a join
on the exercise collection to provide up to date data.
The newly updated PUT/PATCH functions were not doing this,
so the Records they returned had stale exercise data
(noticeable in set headers not saving display fields with
split weight).