change back to CSR
SSR may be faster on initial page load, but it has to load from server on
every route. It also introduces a delay on route (when the server is loading).
This makes loading seem like lag/delay, and is a worse experience than having a
potentially longer, but visible,
CSR loading transition.