Enable cssMode on outer swiper
This should be the last time fiddling with this.
History swipers actually work fine without cssMode
when limiting the max records pulled.
May want to lean down the outer swiper at some point.
Also: I remembered that the reason there are a bunch of duplicate records
on that one day was due to a bug with copying sessions. It must have actually
saved those extra sessions, despite crashing client side.