move history into record card
Not sure about keeping it this way, but see how it goes.
It makes logical sense to keep the history swiper visually inside the card,
since it IS a part of the card component and is linked to the record.
The main thing is just that the loading states look kind of weird and
there's no title for loaded data saying it's the history