fix unlimited history query bug
turns out on init each record was requesting the entire record table
becuse the filter was an empty object.
Changed to init to undefined and default to a limit if none is provided.
This should drastically decrease load time for fetching records,
and may be the cause of the strange add new records behavior