update forms to not use state/useEffects, refactor ExerciseForm
ExerciseForm had been using a copy of the exercise via useGuaranteedExercise
for mutations to avoid rerendering the form on change,
which would reset focus and dropdowns.
With the new simplification this is no longer necessary, so it
can share the same pattern as the other forms.
Also move useNames to be a return for useSwr objects
(memoizing doesn't do anything since the dep is an array)