add targetDate to determine loading state
Delay seems to not be inherent to the router. Deleting swiper component
and just showing a div with the date shows the date update with no delay.
Likely due to server side rendering taking too long for the larger component.

Tried changing targetDate to the actual date and it is faster, but starts
delaying when clicking the calendar. Thinking the root cause may simply be
loading too much into the dom all at once.

TargetDate may be slower but it feels more reactive because of the loading spinner.