remove editable history filter
With the addition of setTypes that seems not needed.
It also seems confusing having the filter icon at the top of the record instead
of with the actual history, but there isn't a great way to colocate them visually.
Colocation would also be good to encapsulate state to the history swiper.

The idea now is that record cards just show the history matching
their set type / modifiers, and there should be a
separate history page that allows arbitrary querying.
The separate page could also be linked by a record action
(like Manage Exercise) that links to
the history page with the current record's fields filled in.