memoize record mutation functions
This was tricky to figure out.
Children can't ever access the full record because mutate() always
triggers a rerender for anything using the record.
Memoized functions must also not access the full record or the memo will
update too often. Have to put everything within functions that accept callbacks
which pass in the current record.

End result is that useCurrentRecord / recordContext are not very useful since
they can't actually be used in the children components without triggering
expensive rerenders.