split RecordQuery into required and partial types
Has been tricky to iron out set type because records require a set type,
but the query api has all fields as optional.
This solution makes the query required on the frontend with a separate type.

Also renames ArrayMatchType to MatchType. SetType can use the same nomenclature,
though the effect is different.
An exact match matches the set type, and a partial match matches
any sets that would fit in that set type.