filter record activeModifiers based on exercise modifiers
When a user removes a modifier from an exercise, that modifier will still be in any records that use it.
This is a problem because you can't filter history for those records since you can't select that modifier anymore,
and if you're removing the modifier you are signifying you don't want to track it anymore.
So when fetching from mongo we ignore those modifiers,
but still maintain the data in case the user re-enables the modifier later.