show all exercises, not just active
on reflection, "status" seems not particularly useful. In practice it has only been used on exercises.
Exercises represent a record of work done, whereas
modifiers/categories are for classification/tracking.
It's not like you'll ever need to archive categories like "abs" or "chest". In fact maybe there isn't really much value in user maintained categories.
Or, there could be a default set of categories that the user could add to.

This commit allows for archived exercises to still be visible in selectors. This gives them a visible distinction and moves them behind "active" exercises, so they still are moved out of the way of active exercises
but there aren't issues with having a previous record with a now archived exercise that no longer appears in the selector.