remove status from modifier and category
Exercise, modifier, and category are all joined by a common AsyncSelectorOption interface, which presumes they have a status.
AsyncSelector depends on the status to display "add new item" text.
By isolating the groupBy status to ExerciseSelector we hide the underlying dependency on a status.
Modifier and Category secretly have a permanent status of "active". This also leaves the possibility that it will be easy to re-add status to categories/modifiers in the future if desired.