memoize equipment weight field
We don't want to do a LoadedExerciseForm like LoadedRecordCard because we already have all the data available from the exercise selector.
The selector needs the full exercise objects because it looks at the name and the status.
Even if we had a LoadedExerciseForm, it would still need a full Exercise as a prop, so it would still rerender whenever any field changes.

So instead of trying to memoize the whole form, we need to memoize the fields individually so they only rerender if their relevant data changes.

The big hurdle is the "handleUpdate" function, which was solved by isolating the old exercise data to within the mutate() function rather than passing it in as an arg.