fix styling in SetTypeSelect for standard variant; fix swiper preventing opening select
the select Icon is slightly off center when using standard variant when compared with autocompletes.

Also fixes onOpen not triggering due to swiper intercepting clicks looking for swipes. Swiper has special handling for buttons, so we have to treat the select text as a button and manually trigger handleOpen on click