update to new slotProps format in anticipation of v7
InputProps, SelectProps etc are all deprecated now.

As a bonus, there is no longer a need for the hacky style tweaks "fixSelectBackground" and "standardVariantSx"

On the other hand, slotProps makes some of this code a lot more cumbersome and it doesn't provide the right typing info to each slot field.