refactor to throw exception on 404 api res
we are asserting that the res json is always type T. This was not true -- it could be null if the api returned 404.
Refactored to throw an exception in the case of 404.
Previous comment about an error being harder to handle was out of date and not relevant. All usages of fetchJson expect a response of type T and assume it is not null.

Also, verbose logs were removed for being redundant.