fix copy session not triggering rerender
copy session was not triggering a rerender when copied from an existing empty session because the copy function was mutating the existing session's records instead of making a new record array.
This prevented react from recognizing there was a change, and it would fail to rerender. Clicking the button would have no visible effect until something else triggered a rerender.