update npm pkgs
NOT updating zod to v4. Zod v4 introduces a number of breaking changes, and changes the behavior of a lot of its parsing logic.
With the eventual planned transition to nextjs app router zod may become redundant, so it's not worth tinkering with anymore to iron out the new validation schemas.