update CI to build and start prod app to fix flakiness
with app router the test server keeps doing HMR reloads in CI which reloads the page in the middle of tests and causes them to fail.
This does not happen locally, but consistently does in CI.

Note: must remove .env.production, checking this in because it works. CI is inexplicably hanging permanently using other methods