disable lint
In App router, nextjs attempts static generation (SSG) of all pages, including using a connection to live data.
This can be disabled with next build --experimental-build-mode compile,
but it's not really vital to build without local mongo running and the experimental option is not stable / may introduce oddities to the app.
SSG does nothing for the app since everything is behind a login, so nothing can be staticly generated.