fix build failure missing suspense
The suspense doesn't seem to actually be doing anything but react is looking for a suspense and errors if there is not one.
Unsure what exactly needs the suspense. Maybe something with prefetching using tanstack.
The suspense should not actually be triggered ever since the server components are rendered with SSR so there is never an undefined/loading state.