switch to bun pkg manager
bun can be installed with npm. It can be used as a full replacement for node, but it does not yet have 100% functionality of node and does not work with the app. When running with "bun" instead of "npm", it is still using node under the hood if the script has something like #!/usr/bin/env node at the top. To force using bun runtime use the flag "--bun", but the app cannot currently run with bun runtime.

We can use bun just as a package manager though, which has a few minor advantages:
-  package.json scripts only need "bun dev" vs "npm run dev"
- "bun install" may potentially be faster
- bun.lock is a smaller file than package-lock.json
- mongo scripts can use bun runtime and do run noticeably faster
- don't need cross-env; bun automatically pulls in the env vars