fix prefetch not working properly
prefetching was not actually working because the keys were not the same.
day.endOf('month') was being applied before day.add(1, 'month'), which was making the next month have the date of whatever the last date in the current month was, NOT the actual last date in the next month.
This made the keys not match up if the current month had less days than the next month, so the prefetched data was not used.

Prefetch data now works as expected.