reuse similar querykey; investigate syncing across tabs
mainly was investigating how to sync across tab so if "manage" was open and in another tab an exercise was added / removed the manage page would update the delete button.
BUT it turns out tanstack does not sync across tabs. There is an extra experimental lib you can add to enable that (broadcast).
Will not add at this time since that's not a big concern. Typical use is in a single tab, and if you make a mutation on an outdated cache there will just be an error saying to reload