do not hide modifiers that have been removed from exercises
If a record has a modifier that later is removed from the exercise, it has been hiding it. Now just continues to show it. Could cause confusing behavior previously with hiding modifiers that still actually exist. Now you would see if the record has an old modifier on it, though it may be harder to find history.